requirements.txt
flask
python-dotenv
flask-wtf
flask-admin[images,export,sqlalchemy]
flask-login
flask-sqlalchemy
flask-alembic
flask-mail